+++ ThinkPHP/Library/Think/Db/Driver/Pgsql.class.php
@@ -8,214 +8,47 @@
 // +----------------------------------------------------------------------
 // | Author: liu21st <liu21st@gmail.com>
 // +----------------------------------------------------------------------
+
 namespace Think\Db\Driver;
-use Think\Db;
-defined('THINK_PATH') or exit();
+use Think\Db\Driver;
+
 /**
  * Pgsql数据库驱动
  */
-class Pgsql extends Db{
+class Pgsql extends Driver{
 
     /**
-     * 架构函数 读取数据库配置信息
+     * 解析pdo连接的dsn信息
      * @access public
-     * @param array $config 数据库配置数组
+     * @param array $config 连接信息
+     * @return string
      */
-    public function __construct($config='') {
-        if ( !extension_loaded('pgsql') ) {
-            E(L('_NOT_SUPPORT_').':pgsql');
+    protected function parseDsn($config){
+        $dsn  =   'pgsql:dbname='.$config['database'].';host='.$config['hostname'];
+        if(!empty($config['hostport'])) {
+            $dsn  .= ';port='.$config['hostport'];
         }
-        if(!empty($config)) {
-            $this->config   =   $config;
-            if(empty($this->config['params'])) {
-                $this->config['params'] =   array();
-            }
-        }
-    }
-
-    /**
-     * 连接数据库方法
-     * @access public
-     */
-    public function connect($config='',$linkNum=0) {
-        if ( !isset($this->linkID[$linkNum]) ) {
-            if(empty($config))  $config =   $this->config;
-            $pconnect   = !empty($config['params']['persist'])? $config['params']['persist']:$this->pconnect;
-            $conn = $pconnect ? 'pg_pconnect':'pg_connect';
-            $this->linkID[$linkNum] =  $conn('host='.$config['hostname'].' port='.$config['hostport'].' dbname='.$config['database'].' user='.$config['username'].'  password='.$config['password']);
-            if (0 !== pg_connection_status($this->linkID[$linkNum])){
-                E($this->error(false));
-            }
-            //设置编码
-            pg_set_client_encoding($this->linkID[$linkNum], $config['charset']);
-        }
-        return $this->linkID[$linkNum];
-    }
-
-    /**
-     * 释放查询结果
-     * @access public
-     */
-    public function free() {
-        pg_free_result($this->queryID);
-        $this->queryID = null;
-    }
-
-    /**
-     * 执行查询 返回数据集
-     * @access public
-     * @param string $str  sql指令
-     * @return mixed
-     */
-    public function query($str) {
-        $this->initConnect(false);
-        if ( !$this->_linkID ) return false;
-        $this->queryStr = $str;
-        //释放前次的查询结果
-        if ( $this->queryID ) $this->free();
-        N('db_query',1);
-        // 记录开始执行时间
-        G('queryStartTime');
-        $this->queryID = pg_query($this->_linkID,$str);
-        $this->debug();
-        if ( false === $this->queryID ) {
-            $this->error();
-            return false;
-        } else {
-            $this->numRows = pg_num_rows($this->queryID);
-            return $this->getAll();
-        }
-    }
-
-    /**
-     * 执行语句
-     * @access public
-     * @param string $str  sql指令
-     * @return integer
-     */
-    public function execute($str) {
-        $this->initConnect(true);
-        if ( !$this->_linkID ) return false;
-        $this->queryStr = $str;
-        //释放前次的查询结果
-        if ( $this->queryID ) $this->free();
-        N('db_write',1);
-        // 记录开始执行时间
-        G('queryStartTime');
-        $result =   pg_query($this->_linkID,$str);
-        $this->debug();
-        if ( false === $result ) {
-            $this->error();
-            return false;
-        } else {
-            $this->numRows = pg_affected_rows($result);
-            $this->lastInsID =   $this->last_insert_id();
-            return $this->numRows;
-        }
-    }
-
-    /**
-     * 用于获取最后插入的ID
-     * @access public
-     * @return integer
-     */
-    public function last_insert_id() {
-        $query  =   "SELECT LASTVAL() AS insert_id";
-        $result =   pg_query($this->_linkID,$query);
-        list($last_insert_id)   =   pg_fetch_array($result,null,PGSQL_ASSOC);
-        pg_free_result($result);
-        return $last_insert_id;
-    }
-
-    /**
-     * 启动事务
-     * @access public
-     * @return void
-     */
-    public function startTrans() {
-        $this->initConnect(true);
-        if ( !$this->_linkID ) return false;
-        //数据rollback 支持
-        if ($this->transTimes == 0) {
-            pg_exec($this->_linkID,'begin;');
-        }
-        $this->transTimes++;
-        return ;
-    }
-
-    /**
-     * 用于非自动提交状态下面的查询提交
-     * @access public
-     * @return boolean
-     */
-    public function commit() {
-        if ($this->transTimes > 0) {
-            $result = pg_exec($this->_linkID,'end;');
-            if(!$result){
-                $this->error();
-                return false;
-            }
-            $this->transTimes = 0;
-        }
-        return true;
-    }
-
-    /**
-     * 事务回滚
-     * @access public
-     * @return boolean
-     */
-    public function rollback() {
-        if ($this->transTimes > 0) {
-            $result = pg_exec($this->_linkID,'abort;');
-            if(!$result){
-                $this->error();
-                return false;
-            }
-            $this->transTimes = 0;
-        }
-        return true;
-    }
-
-    /**
-     * 获得所有的查询数据
-     * @access private
-     * @return array
-     */
-    private function getAll() {
-        //返回数据集
-        $result   =  pg_fetch_all($this->queryID);
-        pg_result_seek($this->queryID,0);
-        return $result;
+        return $dsn;
     }
 
     /**
      * 取得数据表的字段信息
      * @access public
+     * @return array
      */
     public function getFields($tableName) {
-        $result   =  $this->query("select a.attname as \"Field\",
-            t.typname as \"Type\",
-            a.attnotnull as \"Null\",
-            i.indisprimary as \"Key\",
-            d.adsrc as \"Default\"
-            from pg_class c
-            inner join pg_attribute a on a.attrelid = c.oid
-            inner join pg_type t on a.atttypid = t.oid
-            left join pg_attrdef d on a.attrelid=d.adrelid and d.adnum=a.attnum
-            left join pg_index i on a.attnum=ANY(i.indkey) and c.oid = i.indrelid
-            where (c.relname='{$tableName}' or c.relname = lower('{$tableName}'))   AND a.attnum > 0
-                order by a.attnum asc;");
-        $info   =   array();
-        if($result) {
+        list($tableName) = explode(' ', $tableName);
+        $result =   $this->query('select fields_name as "field",fields_type as "type",fields_not_null as "null",fields_key_name as "key",fields_default as "default",fields_default as "extra" from table_msg('.$tableName.');');
+        $info   =   [];
+        if($result){
             foreach ($result as $key => $val) {
-                $info[$val['Field']] = array(
-                'name'    => $val['Field'],
-                'type'    => $val['Type'],
-                'notnull' => (bool) ($val['Null'] == 't'?1:0), // 't' is 'not null'
-                'default' => $val['Default'],
-                'primary' => (strtolower($val['Key']) == 't'),
-                'autoinc' => (strtolower($val['Default']) == "nextval('{$tableName}_id_seq'::regclass)"),
+                $info[$val['field']] = array(
+                    'name'    => $val['field'],
+                    'type'    => $val['type'],
+                    'notnull' => (bool) ($val['null'] === ''), // not null is empty, null is yes
+                    'default' => $val['default'],
+                    'primary' => (strtolower($val['key']) == 'pri'),
+                    'autoinc' => (strtolower($val['extra']) == 'auto_increment'),
                 );
             }
         }
@@ -225,9 +58,10 @@ class Pgsql extends Db{
     /**
      * 取得数据库的表信息
      * @access public
+     * @return array
      */
     public function getTables($dbName='') {
-        $result = $this->query("select tablename as Tables_in_test from pg_tables where  schemaname ='public'");
+        $result =   $this->query("select tablename as Tables_in_test from pg_tables where  schemaname ='public'");
         $info   =   array();
         foreach ($result as $key => $val) {
             $info[$key] = current($val);
@@ -236,44 +70,9 @@ class Pgsql extends Db{
     }
 
     /**
-     * 关闭数据库
-     * @access public
-     */
-    public function close() {
-        if($this->_linkID){
-            pg_close($this->_linkID);
-        }
-        $this->_linkID = null;
-    }
-
-    /**
-     * 数据库错误信息
-     * 并显示当前的SQL语句
-     * @access public
-     * @return string
-     */
-    public function error($result = true) {
-        $this->error = $result?pg_result_error($this->queryID): pg_last_error($this->_linkID);
-        if('' != $this->queryStr){
-            $this->error .= "\n [ SQL语句 ] : ".$this->queryStr;
-        }
-        trace($this->error,'','ERR');
-        return $this->error;
-    }
-
-    /**
-     * SQL指令安全过滤
-     * @access public
-     * @param string $str  SQL指令
-     * @return string
-     */
-    public function escapeString($str) {
-        return pg_escape_string($str);
-    }
-
-    /**
-     * limit
-     * @access public
+     * limit分析
+     * @access protected
+     * @param mixed $lmit
      * @return string
      */
     public function parseLimit($limit) {
@@ -288,4 +87,5 @@ class Pgsql extends Db{
         }
         return $limitStr;
     }
+
 }
