+++ ThinkPHP/Library/Think/Db/Driver/Sqlite.class.php
@@ -8,178 +8,24 @@
 // +----------------------------------------------------------------------
 // | Author: liu21st <liu21st@gmail.com>
 // +----------------------------------------------------------------------
+
 namespace Think\Db\Driver;
-use Think\Db;
-defined('THINK_PATH') or exit();
+use Think\Db\Driver;
+
 /**
  * Sqlite数据库驱动
  */
-class Sqlite extends Db {
+class Sqlite extends Driver {
 
     /**
-     * 架构函数 读取数据库配置信息
+     * 解析pdo连接的dsn信息
      * @access public
-     * @param array $config 数据库配置数组
+     * @param array $config 连接信息
+     * @return string
      */
-    public function __construct($config='') {
-        if ( !extension_loaded('sqlite') ) {
-            E(L('_NOT_SUPPORT_').':sqlite');
-        }
-        if(!empty($config)) {
-            if(!isset($config['mode'])) {
-                $config['mode']	=	0666;
-            }
-            $this->config	=	$config;
-            if(empty($this->config['params'])) {
-                $this->config['params'] =   array();
-            }
-        }
-    }
-
-    /**
-     * 连接数据库方法
-     * @access public
-     */
-    public function connect($config='',$linkNum=0) {
-        if ( !isset($this->linkID[$linkNum]) ) {
-            if(empty($config))	$config	=	$this->config;
-            $pconnect   = !empty($config['params']['persist'])? $config['params']['persist']:$this->pconnect;
-            $conn = $pconnect ? 'sqlite_popen':'sqlite_open';
-            $this->linkID[$linkNum] = $conn($config['database'],$config['mode']);
-            if ( !$this->linkID[$linkNum]) {
-                E(sqlite_error_string(sqlite_last_error($this->linkID[$linkNum])));
-            }
-        }
-        return $this->linkID[$linkNum];
-    }
-
-    /**
-     * 释放查询结果
-     * @access public
-     */
-    public function free() {
-        $this->queryID = null;
-    }
-
-    /**
-     * 执行查询 返回数据集
-     * @access public
-     * @param string $str  sql指令
-     * @return mixed
-     */
-    public function query($str) {
-        $this->initConnect(false);
-        if ( !$this->_linkID ) return false;
-        $this->queryStr = $str;
-        //释放前次的查询结果
-        if ( $this->queryID ) $this->free();
-        N('db_query',1);
-        // 记录开始执行时间
-        G('queryStartTime');
-        $this->queryID = sqlite_query($this->_linkID,$str);
-        $this->debug();
-        if ( false === $this->queryID ) {
-            $this->error();
-            return false;
-        } else {
-            $this->numRows = sqlite_num_rows($this->queryID);
-            return $this->getAll();
-        }
-    }
-
-    /**
-     * 执行语句
-     * @access public
-     * @param string $str  sql指令
-     * @return integer
-     */
-    public function execute($str) {
-        $this->initConnect(true);
-        if ( !$this->_linkID ) return false;
-        $this->queryStr = $str;
-        //释放前次的查询结果
-        if ( $this->queryID ) $this->free();
-        N('db_write',1);
-        // 记录开始执行时间
-        G('queryStartTime');
-        $result	=	sqlite_exec($this->_linkID,$str);
-        $this->debug();
-        if ( false === $result ) {
-            $this->error();
-            return false;
-        } else {
-            $this->numRows = sqlite_changes($this->_linkID);
-            $this->lastInsID = sqlite_last_insert_rowid($this->_linkID);
-            return $this->numRows;
-        }
-    }
-
-    /**
-     * 启动事务
-     * @access public
-     * @return void
-     */
-    public function startTrans() {
-        $this->initConnect(true);
-        if ( !$this->_linkID ) return false;
-        //数据rollback 支持
-        if ($this->transTimes == 0) {
-            sqlite_query($this->_linkID,'BEGIN TRANSACTION');
-        }
-        $this->transTimes++;
-        return ;
-    }
-
-    /**
-     * 用于非自动提交状态下面的查询提交
-     * @access public
-     * @return boolean
-     */
-    public function commit() {
-        if ($this->transTimes > 0) {
-            $result = sqlite_query($this->_linkID,'COMMIT TRANSACTION');
-            if(!$result){
-                $this->error();
-                return false;
-            }
-            $this->transTimes = 0;
-        }
-        return true;
-    }
-
-    /**
-     * 事务回滚
-     * @access public
-     * @return boolean
-     */
-    public function rollback() {
-        if ($this->transTimes > 0) {
-            $result = sqlite_query($this->_linkID,'ROLLBACK TRANSACTION');
-            if(!$result){
-                $this->error();
-                return false;
-            }
-            $this->transTimes = 0;
-        }
-        return true;
-    }
-
-    /**
-     * 获得所有的查询数据
-     * @access private
-     * @return array
-     */
-    private function getAll() {
-        //返回数据集
-        $result = array();
-        if($this->numRows >0) {
-            for($i=0;$i<$this->numRows ;$i++ ){
-                // 返回数组集
-                $result[$i] = sqlite_fetch_array($this->queryID,SQLITE_ASSOC);
-            }
-            sqlite_seek($this->queryID,0);
-        }
-        return $result;
+    protected function parseDsn($config){
+        $dsn  =   'sqlite:'.$config['database'];
+        return $dsn;
     }
 
     /**
@@ -188,17 +34,18 @@ class Sqlite extends Db {
      * @return array
      */
     public function getFields($tableName) {
+        list($tableName) = explode(' ', $tableName);
         $result =   $this->query('PRAGMA table_info( '.$tableName.' )');
         $info   =   array();
         if($result){
             foreach ($result as $key => $val) {
-                $info[$val['Field']] = array(
-                    'name'    => $val['Field'],
-                    'type'    => $val['Type'],
-                    'notnull' => (bool) ($val['Null'] === ''), // not null is empty, null is yes
-                    'default' => $val['Default'],
-                    'primary' => (strtolower($val['Key']) == 'pri'),
-                    'autoinc' => (strtolower($val['Extra']) == 'auto_increment'),
+                $info[$val['field']] = array(
+                    'name'    => $val['field'],
+                    'type'    => $val['type'],
+                    'notnull' => (bool) ($val['null'] === ''), // not null is empty, null is yes
+                    'default' => $val['default'],
+                    'primary' => (strtolower($val['dey']) == 'pri'),
+                    'autoinc' => (strtolower($val['extra']) == 'auto_increment'),
                 );
             }
         }
@@ -222,40 +69,13 @@ class Sqlite extends Db {
     }
 
     /**
-     * 关闭数据库
-     * @access public
-     */
-    public function close() {
-        if ($this->_linkID){
-            sqlite_close($this->_linkID);
-        }
-        $this->_linkID = null;
-    }
-
-    /**
-     * 数据库错误信息
-     * 并显示当前的SQL语句
-     * @access public
-     * @return string
-     */
-    public function error() {
-        $code   =   sqlite_last_error($this->_linkID);
-        $this->error = $code.':'.sqlite_error_string($code);
-        if('' != $this->queryStr){
-            $this->error .= "\n [ SQL语句 ] : ".$this->queryStr;
-        }
-        trace($this->error,'','ERR');
-        return $this->error;
-    }
-
-    /**
      * SQL指令安全过滤
      * @access public
      * @param string $str  SQL指令
      * @return string
      */
     public function escapeString($str) {
-        return sqlite_escape_string($str);
+        return str_ireplace("'", "''", $str);
     }
 
     /**
