+++ ThinkPHP/Library/Think/Db/Driver/Sqlsrv.class.php
@@ -8,187 +8,36 @@
 // +----------------------------------------------------------------------
 // | Author: liu21st <liu21st@gmail.com>
 // +----------------------------------------------------------------------
+
 namespace Think\Db\Driver;
-use Think\Db;
-defined('THINK_PATH') or exit();
+use Think\Db\Driver;
+use PDO;
+
 /**
  * Sqlsrv数据库驱动
  */
-class Sqlsrv extends Db{
+class Sqlsrv extends Driver{
     protected $selectSql  =     'SELECT T1.* FROM (SELECT thinkphp.*, ROW_NUMBER() OVER (%ORDER%) AS ROW_NUMBER FROM (SELECT %DISTINCT% %FIELD% FROM %TABLE%%JOIN%%WHERE%%GROUP%%HAVING%) AS thinkphp) AS T1 %LIMIT%%COMMENT%';
-    /**
-     * 架构函数 读取数据库配置信息
-     * @access public
-     * @param array $config 数据库配置数组
-     */
-    public function __construct($config='') {
-        if ( !function_exists('sqlsrv_connect') ) {
-            E(L('_NOT_SUPPORT_').':sqlsrv');
-        }
-        if(!empty($config)) {
-            $this->config	=	$config;
-        }
-    }
+    // PDO连接参数
+    protected $options = array(
+        PDO::ATTR_CASE              =>  PDO::CASE_LOWER,
+        PDO::ATTR_ERRMODE           =>  PDO::ERRMODE_EXCEPTION,
+        PDO::ATTR_STRINGIFY_FETCHES =>  false,
+        PDO::SQLSRV_ATTR_ENCODING   =>  PDO::SQLSRV_ENCODING_UTF8,
+    );
 
     /**
-     * 连接数据库方法
+     * 解析pdo连接的dsn信息
      * @access public
+     * @param array $config 连接信息
+     * @return string
      */
-    public function connect($config='',$linkNum=0) {
-        if ( !isset($this->linkID[$linkNum]) ) {
-            if(empty($config))	$config  =  $this->config;
-            $host = $config['hostname'].($config['hostport']?",{$config['hostport']}":'');
-            $connectInfo  =  array('Database'=>$config['database'],'UID'=>$config['username'],'PWD'=>$config['password'],'CharacterSet' => C('DEFAULT_CHARSET'));
-            $this->linkID[$linkNum] = sqlsrv_connect( $host, $connectInfo);
-            if ( !$this->linkID[$linkNum] )  $this->error(false);
+    protected function parseDsn($config){
+        $dsn  =   'sqlsrv:dbname='.$config['database'].';Server='.$config['hostname'];
+        if(!empty($config['hostport'])) {
+            $dsn  .= ','.$config['hostport'];
         }
-        return $this->linkID[$linkNum];
-    }
-
-    /**
-     * 释放查询结果
-     * @access public
-     */
-    public function free() {
-        sqlsrv_free_stmt($this->queryID);
-        $this->queryID = null;
-    }
-
-    /**
-     * 执行查询  返回数据集
-     * @access public
-     * @param string $str  sql指令
-     * @param array $bind 参数绑定
-     * @return mixed
-     */
-    public function query($str,$bind=array()) {
-        $this->initConnect(false);
-        if ( !$this->_linkID ) return false;
-        //释放前次的查询结果
-        if ( $this->queryID ) $this->free();
-        N('db_query',1);
-        // 记录开始执行时间
-        G('queryStartTime');
-        $str    =   str_replace(array_keys($bind),'?',$str);
-        $bind   =   array_values($bind);
-        $this->queryStr = $str;
-        $this->queryID = sqlsrv_query($this->_linkID,$str,$bind, array( "Scrollable" => SQLSRV_CURSOR_KEYSET));
-        $this->debug();
-        if ( false === $this->queryID ) {
-            $this->error();
-            return false;
-        } else {
-            $this->numRows = sqlsrv_num_rows($this->queryID);
-            return $this->getAll();
-        }
-    }
-
-    /**
-     * 执行语句
-     * @access public
-     * @param string $str  sql指令
-     * @param array $bind 参数绑定
-     * @return integer
-     */
-    public function execute($str,$bind=array()) {
-        $this->initConnect(true);
-        if ( !$this->_linkID ) return false;
-        //释放前次的查询结果
-        if ( $this->queryID ) $this->free();
-        N('db_write',1);
-        // 记录开始执行时间
-        G('queryStartTime');
-        $str    =   str_replace(array_keys($bind),'?',$str);
-        $bind   =   array_values($bind);
-        $this->queryStr = $str;
-        $this->queryID=	sqlsrv_query($this->_linkID,$str,$bind);
-        $this->debug();
-        if ( false === $this->queryID ) {
-            $this->error();
-            return false;
-        } else {
-            $this->numRows = sqlsrv_rows_affected($this->queryID);
-            $this->lastInsID = $this->mssql_insert_id();
-            return $this->numRows;
-        }
-    }
-
-    /**
-     * 用于获取最后插入的ID
-     * @access public
-     * @return integer
-     */
-    public function mssql_insert_id() {
-        $query  =   "SELECT @@IDENTITY as last_insert_id";
-        $result =   sqlsrv_query($this->_linkID,$query);
-        list($last_insert_id)   =   sqlsrv_fetch_array($result);
-        sqlsrv_free_stmt($result);
-        return $last_insert_id;
-    }
-
-    /**
-     * 启动事务
-     * @access public
-     * @return void
-     */
-    public function startTrans() {
-        $this->initConnect(true);
-        if ( !$this->_linkID ) return false;
-        //数据rollback 支持
-        if ($this->transTimes == 0) {
-            sqlsrv_begin_transaction($this->_linkID);
-        }
-        $this->transTimes++;
-        return ;
-    }
-
-    /**
-     * 用于非自动提交状态下面的查询提交
-     * @access public
-     * @return boolean
-     */
-    public function commit() {
-        if ($this->transTimes > 0) {
-            $result = sqlsrv_commit($this->_linkID);
-            $this->transTimes = 0;
-            if(!$result){
-                $this->error();
-                return false;
-            }
-        }
-        return true;
-    }
-
-    /**
-     * 事务回滚
-     * @access public
-     * @return boolean
-     */
-    public function rollback() {
-        if ($this->transTimes > 0) {
-            $result = sqlsrv_rollback($this->_linkID);
-            $this->transTimes = 0;
-            if(!$result){
-                $this->error();
-                return false;
-            }
-        }
-        return true;
-    }
-
-    /**
-     * 获得所有的查询数据
-     * @access private
-     * @return array
-     */
-    private function getAll() {
-        //返回数据集
-        $result = array();
-        if($this->numRows >0) {
-            while($row = sqlsrv_fetch_array($this->queryID,SQLSRV_FETCH_ASSOC))
-                $result[]   =   $row;
-        }
-        return $result;
+        return $dsn;
     }
 
     /**
@@ -197,16 +46,15 @@ class Sqlsrv extends Db{
      * @return array
      */
     public function getFields($tableName) {
-        $result = $this->query("
-            SELECT column_name,data_type,column_default,is_nullable
-            FROM   information_schema.tables AS t
-            JOIN   information_schema.columns AS c
-            ON     t.table_catalog = c.table_catalog
-            AND    t.table_schema  = c.table_schema
-            AND    t.table_name    = c.table_name
-            WHERE  t.table_name = '{$tableName}'");
-        $pk = $this->query("SELECT * FROM INFORMATION_SCHEMA.KEY_COLUMN_USAGE WHERE TABLE_NAME='{$tableName}'");
-        $info = array();
+        list($tableName) = explode(' ', $tableName);
+        $result =   $this->query("SELECT   column_name,   data_type,   column_default,   is_nullable
+        FROM    information_schema.tables AS t
+        JOIN    information_schema.columns AS c
+        ON  t.table_catalog = c.table_catalog
+        AND t.table_schema  = c.table_schema
+        AND t.table_name    = c.table_name
+        WHERE   t.table_name = '$tableName'");
+        $info   =   array();
         if($result) {
             foreach ($result as $key => $val) {
                 $info[$val['column_name']] = array(
@@ -214,7 +62,7 @@ class Sqlsrv extends Db{
                     'type'    => $val['data_type'],
                     'notnull' => (bool) ($val['is_nullable'] === ''), // not null is empty, null is yes
                     'default' => $val['column_default'],
-                    'primary' => $val['column_name'] == $pk[0]['COLUMN_NAME'],
+                    'primary' => false,
                     'autoinc' => false,
                 );
             }
@@ -262,7 +110,7 @@ class Sqlsrv extends Db{
         }
         return $key;   
     }
-    
+
     /**
      * limit
      * @access public
@@ -313,33 +161,4 @@ class Sqlsrv extends Db{
         return $this->execute($sql,$this->parseBind(!empty($options['bind'])?$options['bind']:array()));
     }
 
-    /**
-     * 关闭数据库
-     * @access public
-     */
-    public function close() {
-        if ($this->_linkID){
-            sqlsrv_close($this->_linkID);
-        }
-        $this->_linkID = null;
-    }
-
-    /**
-     * 数据库错误信息
-     * 并显示当前的SQL语句
-     * @access public
-     * @return string
-     */
-    public function error($result = true) {
-        $errors = sqlsrv_errors();
-        $this->error    =   '';
-        foreach( $errors as $error ) {
-            $this->error .= $error['code'].':'.$error['message'];
-        }
-        if('' != $this->queryStr){
-            $this->error .= "\n [ SQL语句 ] : ".$this->queryStr;
-        }
-        $result? trace($this->error,'','ERR'):E($this->error);
-        return $this->error;
-    }
 }
